Unifies command line interface with fortls files
The two interfaces now share all options which should decrease the
maintainance time and also make easier adding new options.

A dictionary with the possible command line options is passed into
LangServer which is then used to set class attributes.

A unittest has been added which tests both the command line interface and
the LangServer configuration file loading functions

This commit also deprecates the `--preserve_keyword_order` argument
and replaces it with its opposite `--sort_keywords`. Because fortls
should not by default be altering the display of the written code.