Adds unittest for autocompletion as snippets
Usually language servers do not treat autocomplete like this, they are
configured to register only the name of the function and then provide
signature help when completing (manually) the argument list.

I suspect that editors with limited UI support would not be able to have
that so instead a more traditional `Tab` completion is available.