Added fortls autoupdate during initialisation
The server will ping PyPi and check if a version is available greater
than the currently installed version. If an old version is detected
fortls will download the new version to the pip localtion of the
current Python distribution.

This should allow for fortls to be correctly installed most times.
Obviously if running from a directory structure that does not
adhere to how pip/conda install packages this might not work.