Cast final animations to Path objects
The fonts are not portable, hence I can either edit
the SVG to embed the fonts or I can cast to path all text objects.

The workbook still contains the objects with their
actual fonts so if an edit needs to be made in the final
SVGs they will have to be re-created from the main workbook.