Ignore preprocessor lines when detecting fixed form fortran
Currently a file with the content

```
#if defined(A) && !defined(B)
c This is a fixed-form style comment
#endif
```

is categorized as free-form Fortran because the preprocess if-clause contains ampersands and exclamation marks.

This commit makes the detection algorithm ignore all lines starting with `#` if pre-processing is enabled for that file. Pre-processor lines are not valid Fortran code and should IMO not be taken into account when trying to determine the Fortran form.