refactor: remove preprocessor whitespace test
This test is problematic for a few reasons and is in need of refactoring. A summary of the issues are:
- macro expansions are already covered in another test
- test too big
- it pollutes the global workspace of symbols
- names of objects are too generic
- tests indirectly the preprocessor

Its purpose in this PR is to test the preprocessor REGEX patterns work
with leading whitespaces and can perform macro expansions.
I have opted to remove and write a more compact test.